fix transformers and quantization imports (#1770)
* fix transformers and quantization imports

* update GHA

* transformers dependency

* move import

* moving tests around

* fix torch version issue

* move all modifiers under pytorch

* fixing 1.9 tests

* fix key error

* bug fix for 1.9

* move torch imports out of base quant
