+++ src/sparseml/modifiers/quantization/__init__.py
@@ -15,4 +15,3 @@
 # flake8: noqa
 
 from .base import *
-from .pytorch import *
