+++ src/sparseml/modifiers/quantization/base.py
@@ -15,10 +15,6 @@
 from typing import Any, Dict, List, Optional
 
 from sparseml.core import Event, Modifier, State
-from sparseml.modifiers.quantization.utils.quantization_scheme import (
-    QuantizationScheme,
-    QuantizationSchemeLoadable,
-)
 
 
 __all__ = ["QuantizationModifier"]
@@ -57,15 +53,6 @@ class QuantizationModifier(Modifier):
     |       model_fuse_fn_name: 'fuse_module'
     |       strict: True
 
-    :param scheme: Default QuantizationScheme to use when enabling quantization
-        in a module. May also be a dictionary to be loaded into the QuantizationScheme
-        class. A string alias may also be used, supported aliases:
-        ['default', 'deepsparse', 'tensorrt'].
-        If None, the default scheme (`QuantizationScheme()`) will be used.
-        Default is None
-    :param scheme_overrides: optional mapping of module type names or submodule type
-        names to quantization schemes to override them with. If a scheme is mapped to
-        'default', then it will use the scheme set in the modifier scheme property
     :param ignore: optional list of module class names or submodule names
         to not quantize. Default is None
     :param disable_quantization_observer_epoch: Epoch to disable updates to the module
@@ -85,8 +72,6 @@ class QuantizationModifier(Modifier):
         scheme_overrides or ignore are not found in a given module. Default True
     """
 
-    scheme: Optional[QuantizationSchemeLoadable] = None
-    scheme_overrides: Optional[Dict[str, QuantizationSchemeLoadable]] = None
     ignore: Optional[List[str]] = None
     disable_quantization_observer_epoch: Optional[float] = None
     freeze_bn_stats_epoch: Optional[float] = None
@@ -98,10 +83,6 @@ class QuantizationModifier(Modifier):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.scheme = QuantizationScheme.load(self.scheme)
-        self.scheme_overrides = _load_quantization_schemes_dict(
-            self.scheme_overrides, self.scheme
-        )
         if self.model_fuse_fn_kwargs is None:
             self.model_fuse_fn_kwargs = {}
         if self.ignore is None:
@@ -158,24 +139,3 @@ class QuantizationModifier(Modifier):
 
     def on_initialize_structure(self, state: State, **kwargs):
         pass  # nothing needed for this modifier
-
-
-class _QuantizationSchemesDict(dict):
-    # wrapper class for dict to override the __str__ method for yaml serialization
-
-    def __str__(self):
-        return str({submodule: scheme.dict() for submodule, scheme in self.items()})
-
-
-def _load_quantization_schemes_dict(
-    schemes_dict: Optional[Dict[str, QuantizationSchemeLoadable]],
-    default_scheme: QuantizationScheme,
-) -> Dict[str, QuantizationScheme]:
-    if schemes_dict is None:
-        return {}
-    return _QuantizationSchemesDict(
-        {
-            submodule: QuantizationScheme.load(scheme, default=default_scheme)
-            for submodule, scheme in schemes_dict.items()
-        }
-    )
